move dht related classes to lbrynet.dht
