+++ lbrynet/tests/functional/test_streamify.py
@@ -13,7 +13,7 @@ from lbrynet.core.StreamDescriptor import StreamDescriptorIdentifier
 from lbrynet.file_manager.EncryptedFileCreator import create_lbry_file
 from lbrynet.lbry_file.client.EncryptedFileOptions import add_lbry_file_to_sd_identifier
 from lbrynet.core.StreamDescriptor import get_sd_info
-from lbrynet.core.PeerManager import PeerManager
+from lbrynet.dht.peermanager import PeerManager
 from lbrynet.core.RateLimiter import DummyRateLimiter
 
 from lbrynet.tests import mocks
